feat: move timer to server
