+++ app-dev/cloud-quiz/app/api/start-game/route.ts
@@ -1,7 +1,7 @@
 import { db } from '@/app/lib/firebase-server-initialization';
 import { getAuthenticatedUser } from '@/app/lib/server-side-auth'
 import { gameStates } from '@/app/types';
-import { FieldValue } from 'firebase-admin/firestore'; 
+import { FieldValue } from 'firebase-admin/firestore';
 import { NextRequest, NextResponse } from 'next/server'
 
 export async function POST(request: NextRequest) {
@@ -21,8 +21,8 @@ export async function POST(request: NextRequest) {
 
   // Validate request
   const { gameId } = await request.json();
-  
-  if(!gameId) {
+
+  if (!gameId) {
     // Respond with JSON indicating an error message
     return new NextResponse(
       JSON.stringify({ success: false, message: 'no game id provided' }),
@@ -33,8 +33,8 @@ export async function POST(request: NextRequest) {
   const gameRef = await db.collection("games").doc(gameId);
   const gameDoc = await gameRef.get();
   const game = gameDoc.data()
-  
-  if(game.leader.uid !== authUser.uid) {
+
+  if (game.leader.uid !== authUser.uid) {
     // Respond with JSON indicating an error message
     return new NextResponse(
       JSON.stringify({ success: false, message: 'no game found' }),
@@ -46,7 +46,34 @@ export async function POST(request: NextRequest) {
   await gameRef.update({
     state: gameStates.AWAITING_PLAYER_ANSWERS,
     startTime: FieldValue.serverTimestamp(),
+    currentTimerStart: FieldValue.serverTimestamp(),
   });
+
+  // start automatic question progression
+  async function onQuestionTimeExpired() {
+    await gameRef.update({
+      state: gameStates.SHOWING_CORRECT_ANSWERS,
+    });
+    setTimeout(onAnswerTimeExpired, game.timePerAnswer * 1000)
+  }
+
+  async function onAnswerTimeExpired() {
+    const gameDoc = await gameRef.get();
+    const game = gameDoc.data();
+    if (game.currentQuestionIndex < Object.keys(game.questions).length - 1) {
+      await gameRef.update({
+        state: gameStates.AWAITING_PLAYER_ANSWERS,
+        currentQuestionIndex: game.currentQuestionIndex + 1,
+      });
+      setTimeout(onQuestionTimeExpired, game.timePerQuestion * 1000)
+    } else {
+      await gameRef.update({
+        state: gameStates.GAME_OVER,
+      });
+    }
+  }
+
+  setTimeout(onQuestionTimeExpired, game.timePerQuestion * 1000)
 
   return NextResponse.json('successfully started game', { status: 200 })
 }
